section-services: Add section and list component
