+++ src/app/(site)/page.tsx
@@ -7,8 +7,6 @@ import { Logo } from '@/components/ui/logo/logo';
 
 import { Socials } from '@/components/ui/socials/socials';
 import { useEffect, useState } from 'react';
-import { getPolitics, getServices } from '@/../sanity/api';
-import { ServiceCard } from '@/components/common/service-card/service-card';
 
 import { NavMenu } from '@/components/ui/nav-menu/nav-menu';
 import { Accordion } from '@/components/common/accordion/accordion';
@@ -22,32 +20,23 @@ import therapyData from '@/data/therapy.json';
 import { TherapyItem } from '@/components/common/therapy-item/therapy-item';
 import { Modal } from '@/components/ui/modal';
 import FormFeedbackMessage from '@/components/common/form-feedback-message/form-feedback-message';
+import { getPolitics } from '@/../sanity/api';
 
 import { TargetAudience } from '@/sections/target-audience/target-audience';
 
 import { PortableText } from '@portabletext/react';
 
 import { ContactForm } from '@/components/common/contact-form';
-
-type Service = {
-  _key: string;
-  title: string;
-};
+import { Services } from '@/sections/services/services';
 
 const Home = () => {
-  const [services, setServices] = useState<Service[]>([]);
   const [isOpen, setIsOpen] = useState(false);
   const [status, setStatus] = useState<'success' | 'failed'>('success');
   const [politics, setPolitics] = useState([]);
 
+  console.log('politics:', politics);
+
   useEffect(() => {
-    async function fetchData() {
-      const data = await getServices();
-      setServices(data);
-    }
-
-    fetchData();
-
     async function fetchDataPolitics() {
       const data = await getPolitics();
       setPolitics(data);
@@ -56,11 +45,9 @@ const Home = () => {
     fetchDataPolitics();
   }, []);
 
-  console.log('services:', services);
-  console.log('politics:', politics);
-
   return (
     <main>
+      <Services />
       <TargetAudience />
       <div className="container">
         <Accordion />
@@ -97,11 +84,6 @@ const Home = () => {
           <NavMenu section="footer" />
         </div>
 
-        <ul className="flex flex-col md:flex-row flex-wrap gap-8 md:gap-x-[94px] md:gap-y-[60px] xl:gap-x-[125px]">
-          {services.map((service) => (
-            <ServiceCard key={service._key} title={service.title} />
-          ))}
-        </ul>
         <ul>
           {reviews.reviews.map((review) => (
             <ReviewCard
