+++ src/components/common/services-list/types.ts
@@ -1,0 +1,4 @@
+export type ServiceData = {
+  _key: string;
+  title: string;
+};
